Crate rp2040_boot2

Statics

BOOT_LOADER_AT25SF128A

The bootloader to use if you want to boot from an AT25SF128A flash device

BOOT_LOADER_GD25Q64CS

The bootloader to use if you want to boot from an GD25Q64CS flash device

BOOT_LOADER_GENERIC_03H

The bootloader to use if you want to boot from a generic flash device

BOOT_LOADER_IS25LP080

The bootloader to use if you want to boot from an IS25LP080 flash device

BOOT_LOADER_RAM_MEMCPY

The bootloader to use if you want to copy code to RAM and then boot from RAM

BOOT_LOADER_W25Q080

The bootloader to use if you have a W25Q080 flash device

BOOT_LOADER_W25X10CL

The bootloader to use if you want to boot from an W25X10CL flash device